OAuth/requirements.txt
flask
flask-jwt-extended
python-dotenv
requests
fastapi[standard]
python-jose[cryptography]